+++ src/configurations/RealVectorSpaceConfig.cpp
@@ -4,5 +4,5 @@
 
 #include "RealVectorSpaceConfig.h"
 
-int RealVectorSpaceConfig::NUM_INTERPOLATION_VALIDITY_CHECKS    = 15;
+size_t RealVectorSpaceConfig::NUM_INTERPOLATION_VALIDITY_CHECKS = 15;
 float RealVectorSpaceConfig::EQUALITY_THRESHOLD                 = 1e-4;
